engine/requirements.txt
nba_api
basketball_reference_web_scraper
psycopg2-binary
redis
pandas
numpy
python-dotenv
schedule
click
